+++ src/main/kotlin/com/yourssu/scouter/ats/application/domain/applicant/ApplicantController.kt
@@ -2,10 +2,6 @@ package com.yourssu.scouter.ats.application.domain.applicant
 
 import com.yourssu.scouter.ats.business.domain.applicant.ApplicantDto
 import com.yourssu.scouter.ats.business.domain.applicant.ApplicantService
-import com.yourssu.scouter.ats.business.domain.applicant.ApplicantSyncResult
-import com.yourssu.scouter.ats.business.domain.applicant.ApplicantSyncService
-import com.yourssu.scouter.common.application.support.authentication.AuthUser
-import com.yourssu.scouter.common.application.support.authentication.AuthUserInfo
 import jakarta.validation.Valid
 import java.net.URI
 import org.springframework.http.ResponseEntity
@@ -21,7 +17,6 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 class ApplicantController(
     private val applicantService: ApplicantService,
-    private val applicantSyncService: ApplicantSyncService,
 ) {
 
     @PostMapping("/applicants")
@@ -34,26 +29,6 @@ class ApplicantController(
         return ResponseEntity.created(URI.create("/applicants/$applicantId")).build()
     }
 
-    @PostMapping("/applicants/include-from-forms")
-    fun includeFromForms(
-        @AuthUser authUserInfo: AuthUserInfo,
-    ): ResponseEntity<ApplicantSyncResponse> {
-        val result: ApplicantSyncResult = applicantSyncService.includeFromForms(authUserInfo.userId)
-        val response: ApplicantSyncResponse = ApplicantSyncResponse.from(result)
-
-        return ResponseEntity.ok(response)
-    }
-
-    @PostMapping("/applicants/include-from-forms/{semesterString}")
-    fun includeFromForms(
-        @AuthUser authUserInfo: AuthUserInfo,
-        @PathVariable semesterString: String,
-    ): ResponseEntity<ApplicantSyncResponse> {
-        val result: ApplicantSyncResult = applicantSyncService.includeFromForms(authUserInfo.userId, semesterString)
-        val response: ApplicantSyncResponse = ApplicantSyncResponse.from(result)
-
-        return ResponseEntity.ok(response)
-    }
 
     @GetMapping("/applicants")
     fun readAll(
